+++ relay-frontend-http/src/server.rs
@@ -1,3 +1,5 @@
+use actix_web::body::MessageBody;
+use actix_web::dev::{ServiceFactory, ServiceRequest, ServiceResponse};
 use actix_web::http::StatusCode;
 use actix_web::middleware::Logger;
 use actix_web::{web, App, HttpResponse, HttpServer};
@@ -20,7 +22,7 @@ struct GetInfo {
 async fn get<BE, T>(data: web::Data<Arc<BE>>, info: web::Path<GetInfo>) -> HttpResponse
 where
     T: Serialize,
-    BE: Backend<T>,
+    BE: Backend<T, T>,
 {
     increment_counter!("http_request", "endpoint" => "get", "queue" => info.queue.clone());
 
@@ -57,7 +59,7 @@ struct ExistsInfo {
 #[tracing::instrument(name = "http_exists", level = "debug", skip_all)]
 async fn exists<BE, T>(data: web::Data<Arc<BE>>, info: web::Path<ExistsInfo>) -> HttpResponse
 where
-    BE: Backend<T>,
+    BE: Backend<T, T>,
 {
     increment_counter!("http_request", "endpoint" => "exists", "queue" => info.queue.clone());
 
@@ -86,9 +88,9 @@ where
 }
 
 #[tracing::instrument(name = "http_enqueue", level = "debug", skip_all)]
-async fn enqueue<BE, T>(data: web::Data<Arc<BE>>, jobs: web::Json<Vec<Job<T>>>) -> HttpResponse
+async fn enqueue<BE, T>(data: web::Data<Arc<BE>>, jobs: web::Json<Vec<Job<T, T>>>) -> HttpResponse
 where
-    BE: Backend<T>,
+    BE: Backend<T, T>,
 {
     increment_counter!("http_request", "endpoint" => "enqueue");
 
@@ -137,7 +139,7 @@ async fn next<BE, T>(
 ) -> HttpResponse
 where
     T: Serialize,
-    BE: Backend<T>,
+    BE: Backend<T, T>,
 {
     increment_counter!("http_request", "endpoint" => "next", "queue" => path_info.queue.clone());
 
@@ -174,7 +176,7 @@ async fn heartbeat<BE, T>(
     state: Option<web::Json<T>>,
 ) -> HttpResponse
 where
-    BE: Backend<T>,
+    BE: Backend<T, T>,
 {
     increment_counter!("http_request", "endpoint" => "heartbeat", "queue" => info.queue.clone());
 
@@ -205,9 +207,9 @@ where
 }
 
 #[tracing::instrument(name = "http_reschedule", level = "debug", skip_all)]
-async fn reschedule<BE, T>(data: web::Data<Arc<BE>>, job: web::Json<Job<T>>) -> HttpResponse
+async fn reschedule<BE, T>(data: web::Data<Arc<BE>>, job: web::Json<Job<T, T>>) -> HttpResponse
 where
-    BE: Backend<T>,
+    BE: Backend<T, T>,
 {
     increment_counter!("http_request", "endpoint" => "reschedule", "queue" => job.0.queue.clone());
 
@@ -242,7 +244,7 @@ struct CompleteInfo {
 #[tracing::instrument(name = "http_delete", level = "debug", skip_all)]
 async fn delete<BE, T>(data: web::Data<Arc<BE>>, info: web::Path<CompleteInfo>) -> HttpResponse
 where
-    BE: Backend<T>,
+    BE: Backend<T, T>,
 {
     increment_counter!("http_request", "endpoint" => "delete", "queue" => info.queue.clone());
 
@@ -288,33 +290,169 @@ impl Server {
     pub async fn run<BE, T>(backend: Arc<BE>, addr: &str) -> anyhow::Result<()>
     where
         T: Serialize + DeserializeOwned + Send + Sync + 'static,
-        BE: Backend<T> + Send + Sync + 'static,
+        BE: Backend<T, T> + Send + Sync + 'static,
     {
-        HttpServer::new(move || {
-            App::new()
-                .app_data(web::Data::new(backend.clone()))
-                .wrap(Logger::new("%a %r %s %Dms"))
-                .route("/v1/queues/jobs", web::post().to(enqueue::<BE, T>))
-                .route("/v1/queues/jobs", web::put().to(reschedule::<BE, T>))
-                .route("/v1/queues/{queue}/jobs", web::get().to(next::<BE, T>))
-                .route(
-                    "/v1/queues/{queue}/jobs/{id}",
-                    web::delete().to(delete::<BE, T>),
-                )
-                .route(
-                    "/v1/queues/{queue}/jobs/{id}",
-                    web::head().to(exists::<BE, T>),
-                )
-                .route("/v1/queues/{queue}/jobs/{id}", web::get().to(get::<BE, T>))
-                .route(
-                    "/v1/queues/{queue}/jobs/{id}",
-                    web::patch().to(heartbeat::<BE, T>),
-                )
-                .route("/health", web::get().to(health))
+        HttpServer::new(move || Self::init_app(backend.clone()))
+            .bind(addr)?
+            .run()
+            .await?;
+        Ok(())
+    }
+
+    // from https://github.com/actix/actix-web/wiki/FAQ#how-can-i-return-app-from-a-function--why-is-appentry-private
+    pub(crate) fn init_app<BE, T>(
+        backend: Arc<BE>,
+    ) -> App<
+        impl ServiceFactory<
+            ServiceRequest,
+            Response = ServiceResponse<impl MessageBody>,
+            Config = (),
+            InitError = (),
+            Error = actix_web::Error,
+        >,
+    >
+    where
+        T: Serialize + DeserializeOwned + Send + Sync + 'static,
+        BE: Backend<T, T> + Send + Sync + 'static,
+    {
+        App::new()
+            .app_data(web::Data::new(backend))
+            .wrap(Logger::new("%a %r %s %Dms"))
+            .route("/v1/queues/jobs", web::post().to(enqueue::<BE, T>))
+            .route("/v1/queues/jobs", web::put().to(reschedule::<BE, T>))
+            .route("/v1/queues/{queue}/jobs", web::get().to(next::<BE, T>))
+            .route(
+                "/v1/queues/{queue}/jobs/{id}",
+                web::delete().to(delete::<BE, T>),
+            )
+            .route(
+                "/v1/queues/{queue}/jobs/{id}",
+                web::head().to(exists::<BE, T>),
+            )
+            .route("/v1/queues/{queue}/jobs/{id}", web::get().to(get::<BE, T>))
+            .route(
+                "/v1/queues/{queue}/jobs/{id}",
+                web::patch().to(heartbeat::<BE, T>),
+            )
+            .route("/health", web::get().to(health))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::client::{Builder as ClientBuilder, Client};
+    use actix_http::HttpService;
+    use actix_http_test::{test_server, TestServer};
+    use actix_service::map_config;
+    use actix_service::ServiceFactoryExt;
+    use actix_web::dev::AppConfig;
+    use chrono::DurationRound;
+    use chrono::Utc;
+    use relay_backend_postgres::PgStore;
+    use uuid::Uuid;
+
+    async fn init_server() -> anyhow::Result<(TestServer, Arc<Client>)> {
+        let db_url = std::env::var("DATABASE_URL")?;
+        let store = Arc::new(PgStore::default(&db_url).await?);
+
+        let srv = test_server(move || {
+            HttpService::build()
+                .h1(map_config(Server::init_app(store.clone()), |_| {
+                    AppConfig::default()
+                }))
+                .tcp()
+                .map_err(|_| ())
         })
-        .bind(addr)?
-        .run()
-        .await?;
+        .await;
+
+        let url = format!("http://{}", srv.addr());
+        let client = ClientBuilder::new(&url).build();
+        Ok((srv, Arc::new(client)))
+    }
+
+    #[tokio::test]
+    async fn test_oneshot() -> anyhow::Result<()> {
+        let (_srv, client) = init_server().await?;
+        let now = Utc::now()
+            .duration_trunc(chrono::Duration::milliseconds(1))
+            .unwrap();
+        let job = Job {
+            id: Uuid::new_v4().to_string(),
+            queue: Uuid::new_v4().to_string(),
+            timeout: 30,
+            max_retries: -1,
+            payload: (),
+            state: None,
+            run_at: Some(now),
+            updated_at: None,
+        };
+
+        client.enqueue(&job).await?;
+
+        let exists = client.exists(&job.queue, &job.id).await?;
+        assert!(exists);
+
+        let mut j = client.get::<(), i32>(&job.queue, &job.id).await?;
+        assert!(j.updated_at.is_some());
+        assert!(j.updated_at.unwrap() >= now);
+        // resetting server side set variable before equality check
+        j.updated_at = None;
+        assert_eq!(job, j);
+
+        let mut jobs = client.poll::<(), i32>(&job.queue, 10).await?;
+        assert_eq!(jobs.len(), 1);
+
+        let mut j2 = jobs.pop().unwrap();
+        j2.updated_at = None;
+        assert_eq!(j2, j);
+
+        client.heartbeat(&job.queue, &job.id, Some(3)).await?;
+
+        let j = client.get::<(), i32>(&job.queue, &job.id).await?;
+        assert_eq!(j.state, Some(3));
+
+        client.remove(&job.queue, &job.id).await?;
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_reschedule() -> anyhow::Result<()> {
+        let (_srv, client) = init_server().await?;
+        let now = Utc::now()
+            .duration_trunc(chrono::Duration::milliseconds(1))
+            .unwrap();
+        let mut job = Job {
+            id: Uuid::new_v4().to_string(),
+            queue: Uuid::new_v4().to_string(),
+            timeout: 30,
+            max_retries: -1,
+            payload: (),
+            state: None,
+            run_at: Some(now),
+            updated_at: None,
+        };
+
+        client.enqueue_batch(&[job.clone()]).await?;
+
+        let mut jobs = client.poll::<(), i32>(&job.queue, 10).await?;
+        assert_eq!(jobs.len(), 1);
+        job = jobs.pop().unwrap();
+
+        job.run_at = Some(
+            Utc::now()
+                .duration_trunc(chrono::Duration::milliseconds(1))
+                .unwrap(),
+        );
+        client.reschedule(&job).await?;
+
+        let mut j = client.get::<(), i32>(&job.queue, &job.id).await?;
+        assert!(j.updated_at.unwrap() >= job.updated_at.unwrap());
+        j.updated_at = None;
+        job.updated_at = None;
+        assert_eq!(j, job);
+
+        client.remove(&job.queue, &job.id).await?;
         Ok(())
     }
 }