+++ CHANGELOG.md
@@ -6,6 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2022-10-29
+### Added
+- HTTP Client.
+- HTTP Consumer abstraction.
+- Worker trait to be used by HTTP Consumer and future frontend consumers.
+- Additional documentation.
+
+### Changed
+- Updated dependencies to latest versions.
+- Updated CI build file dependencies.
+
+### Fixed
+- Generic params for Job struct to two separate as intended.
+- CI build because of openssl and pkg-config issues.
+
 ## [0.9.0] - 2022-09-05
 ### Added
 - Added new exists and get HTTP endpoint to be able to check for a Job's existence or fetch it.
@@ -82,7 +97,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Future Job support using new `run_at` Job field.
 - Reschedule endpoint allowing the Job Runner to manage a unique/singleton Job rescheduling itself.
 
-[Unreleased]: https://github.com/rust-playground/relay-rs/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/rust-playground/relay-rs/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/rust-playground/relay-rs/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/rust-playground/relay-rs/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/rust-playground/relay-rs/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/rust-playground/relay-rs/compare/v0.6.1...v0.7.0